+++ src/Resource/EnvelopeProperties.php
@@ -38,5 +38,9 @@ class EnvelopeProperties extends BaseResource
 
     public string $channelForSender;
 
-    public ?TimePeriod $discardCompletedAfterPeriod;
+    public ?Period $discardCompletedAfterPeriod;
+
+    public Period $expirationPeriod;
+
+    public string $expirationTimeOfDay;
 }